select all post function added

+++ admin/includes/view_all_posts.php
@@ -5,13 +5,15 @@ foreach ($_POST['checkboxArray'] as $postValueId) {
   $bulk_option = $_POST['bulk_option'];
 
   switch ($bulk_option) {
-    case 'published':
+
+      case 'published':
       $query = "UPDATE posts SET post_status = '{$bulk_option}' WHERE post_id = '{$postValueId}' ";
       $upate_to_publish_status = mysqli_query($connection, $query);
       if (!$upate_to_publish_status) {
         die("query failed" . mysql_error($connection));
       }
       break;
+
       case 'draft':
       $query = "UPDATE posts SET post_status = '{$bulk_option}' WHERE post_id = '{$postValueId}' ";
       $upate_to_draft_status = mysqli_query($connection, $query);
@@ -27,8 +29,34 @@ foreach ($_POST['checkboxArray'] as $postValueId) {
         die("query failed" . mysql_error($connection));
       }
       break;
-  }
 
+      case 'clone':
+       $query = "SELECT * FROM posts WHERE post_id = '{$postValueId}' ";
+       $select_post_query = mysqli_query($connection, $query);
+       if (!$select_post_query) {
+         die("query failed" . mysqli_error($connection));
+       }
+       while ($row = mysqli_fetch_assoc($select_post_query)) {
+        $post_id = $row['post_id'];
+        $post_author = $row['post_author'];
+        $post_title = $row['post_title'];
+        $post_category_id = $row['post_catagory_id'];        
+        $post_status = $row['post_status'];
+        $post_tag = $row['post_tag'];
+        $post_comment = $row['post_comment_count'];
+        $post_date = $row['post_date'];
+        $post_image = $row['post_image'];
+        $post_content = $row['post_content'];
+  }
+     
+     $query = "INSERT INTO posts(post_catagory_id, post_title, post_author, post_content, post_date, post_image, post_tag, post_status ) ";
+   $query .= "VALUES ('{$post_category_id}','{$post_title}','{$post_author}','{$post_content}',now(),'{$post_image}','{$post_tag}','{$post_status}' )";
+     $clone_post_query = mysqli_query($connection, $query);
+     if(!$clone_post_query){
+      die("query failed" . mysqli_error($connection));
+     }
+     break;
+}
 }
 }
  ?>
@@ -43,11 +71,12 @@ foreach ($_POST['checkboxArray'] as $postValueId) {
       <option value="published">published</option>
       <option value="draft">Draft</option>
       <option value="delete">Delete</option>
+      <option value="clone">Clone</option>
     </select>
   </div>
   <div class="col-md-4">
     <input type="submit" class="btn btn-success" value="Apply" name="submit">
-    <input type="submit" class="btn btn-primary" value="Add New" >
+    <a href="posts.php?source=add_post" class="btn btn-primary">Add New</a>
   </div>
                             <thead>
                                 <tr>
@@ -62,8 +91,8 @@ foreach ($_POST['checkboxArray'] as $postValueId) {
                                     <th>TAGS</th>
                                     <th>COMMENTS</th>
                                     <th>DATE</th>
-                                    <th></th>
-                                    <th></th>
+                                    <th>EDIT</th>
+                                    <th>DELETE</th>
 
                                  
                                 </tr>
@@ -72,7 +101,7 @@ foreach ($_POST['checkboxArray'] as $postValueId) {
                                 <?php 
                                // query for getting post data from database
                                 
-                                $query = "SELECT * FROM posts";
+                                $query = "SELECT * FROM posts ORDER BY post_id DESC";
                                 $select_post = mysqli_query($connection, $query);
                                 if (!$select_post) {
                                     die("QUERY FAILED" . mysqli_error($select_post));
@@ -120,7 +149,7 @@ foreach ($_POST['checkboxArray'] as $postValueId) {
                                       echo "<td>{$post_comment}</td>";
                                       echo "<td>{$post_date}</td>";
                                       echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";                                      
-                                      echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";                                      
+                                      echo "<td><a onClick=\" javascript: return confirm('are you sure, you want to delete');\" href='posts.php?delete={$post_id}'>Delete</a></td>";                                      
                                       echo "</tr>";
                                 }
